+++ coloraide_extras/__init__.py
@@ -1,5 +1,7 @@
 """ColorAide Extra."""
 from .__meta__ import __version_info__, __version__  # noqa: F401
+from coloraide.util import NaN
+from coloraide.interpolate import Piecewise, Lerp
 from coloraide import Color as Base
 from .spaces.hsi import HSI
 from .spaces.ipt import IPT
@@ -12,7 +14,7 @@ from .spaces.xyy import XyY
 from .spaces.hunter_lab import HunterLab
 from .spaces.prismatic import Prismatic
 
-__all__ = ("Color", "SPACES")
+__all__ = ("Color", "SPACES", 'NaN', 'Piecewise', 'Lerp')
 
 SPACES = [HSI, IPT, IgPgTg, CMY, CMYK, UCS, UVW, XyY, HunterLab, Prismatic]