+++ coloraide_extras/spaces/igpgtg.py
@@ -65,7 +65,7 @@ class IgPgTg(Labish, Space):
     BASE = "xyz-d65"
     NAME = "igpgtg"
     SERIALIZE = ("--igpgtg",)  # type: Tuple[str, ...]
-    CHANNEL_NAMES = ("ig", "tg", "pg")
+    CHANNEL_NAMES = ("ig", "pg", "tg")
     WHITE = WHITES['2deg']['D65']
 
     BOUNDS = (
@@ -87,26 +87,26 @@ class IgPgTg(Labish, Space):
         self._coords[0] = value
 
     @property
-    def tg(self) -> float:
-        """The `Tg` channel."""
-
-        return self._coords[1]
-
-    @tg.setter
-    def tg(self, value: float) -> None:
-        """Set `Tg`."""
-
-        self._coords[1] = value
-
-    @property
     def pg(self) -> float:
         """The `Pg` channel."""
 
-        return self._coords[2]
+        return self._coords[1]
 
     @pg.setter
     def pg(self, value: float) -> None:
         """Set `Pg`."""
+
+        self._coords[1] = value
+
+    @property
+    def tg(self) -> float:
+        """The `Tg` channel."""
+
+        return self._coords[2]
+
+    @tg.setter
+    def tg(self, value: float) -> None:
+        """Set `Tg`."""
 
         self._coords[2] = value